% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_global.R
\name{create_cache}
\alias{create_cache}
\title{Title}
\usage{
create_cache(
  cache_options = list(dir = ft3_options("cache_location"), destroy_on_finalize = TRUE),
  name = "ft_cache",
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{cache_options}{}
}
\description{
Title
}
