% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assignments.R
\name{ft3_assignment_default_settings}
\alias{ft3_assignment_default_settings}
\title{Title}
\usage{
ft3_assignment_default_settings(
  file_ = stop("The file_ argument must be set in assignment_default_settings()."),
  short_title = basename(dirname(file_)),
  title = short_title,
  seed_salt = tools::md5sum(file_),
  data_salt = tools::md5sum(file_),
  hide_before = NA_character_,
  restrict_before = NA_character_,
  fragment = FALSE,
  use_callr = TRUE,
  files = NULL,
  category = ft3_options("default_category_name"),
  sortkey = title,
  on_load = NA_character_,
  exam = FALSE,
  cache = TRUE,
  all_same = TRUE,
  ...
)
}
\arguments{
\item{file_}{}

\item{short_title}{}

\item{title}{}

\item{seed_salt}{}

\item{data_salt}{}

\item{hide_before}{}

\item{restrict_before}{}

\item{use_callr}{}

\item{files}{}

\item{category}{}

\item{sortkey}{}

\item{on_load}{}

\item{exam}{}

\item{...}{}
}
\description{
Title
}
