% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_seed.R
\name{ft3_assignment_seed}
\alias{ft3_assignment_seed}
\title{Title}
\usage{
ft3_assignment_seed(id, seed, data_salt)
}
\arguments{
\item{id}{}

\item{seed}{}

\item{salt}{}

\item{assignment_mode}{}
}
\description{
Title
}
