% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_rmd.R
\name{ft3_assignment_settings_rmd}
\alias{ft3_assignment_settings_rmd}
\title{Get the flextTeaching settings from the yaml header of an assignment}
\usage{
ft3_assignment_settings_rmd(assignment_file)
}
\arguments{
\item{assignment_file}{Path to the assignment source}
}
\description{
Get the flextTeaching settings from the yaml header of an assignment
}
