% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_skeleton.R
\name{ft3_assignment_skeleton}
\alias{ft3_assignment_skeleton}
\title{Title}
\usage{
ft3_assignment_skeleton(
  target_dir,
  title = "Example assignment",
  format = c("html_document", "html_fragment", "pdf_document"),
  ...
)
}
\arguments{
\item{target_dir}{The directory that is to contain the assignments}

\item{title}{Title for the assignment}

\item{format}{One of 'html_document', 'html_fragment', or 'pdf_document'}

\item{...}{Not currently used.}
}
\value{
The path of the file written.
}
\description{
Title
}
