% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_assignment_restrictions.R
\name{ft3_check_assignment_restrictions}
\alias{ft3_check_assignment_restrictions}
\title{Title}
\usage{
ft3_check_assignment_restrictions(settings, pars, check_exam = TRUE)
}
\arguments{
\item{settings}{}

\item{pars}{}
}
\description{
Title
}
