% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg_dir.R
\name{ft3_check_pkg_dir}
\alias{ft3_check_pkg_dir}
\title{Check a package subfolder for flexTeaching3 resources}
\usage{
ft3_check_pkg_dir(pkg)
}
\arguments{
\item{pkg}{}
}
\value{
The normalized support directory path, if valid.
}
\description{
Check a package subfolder for flexTeaching3 resources
}
