% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_doc.R
\name{ft3_compile_doc}
\alias{ft3_compile_doc}
\title{Wrapper to compile assignments of different types}
\usage{
ft3_compile_doc(
  doc_file,
  pars,
  ft3_cache,
  scratch_dir = ft3_options("cache_location"),
  ...
)
}
\arguments{
\item{doc_file}{}

\item{...}{}
}
\description{
Wrapper to compile assignments of different types
}
