% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_rmd.R
\name{ft3_compile_doc_rmd}
\alias{ft3_compile_doc_rmd}
\title{Compile an Rmd flexTeaching assignment}
\usage{
ft3_compile_doc_rmd(rmd_file, pars, local_seed, temp_space_path, ...)
}
\arguments{
\item{rmd_file}{Path to the Rmd assignment file to be compiled}

\item{pars}{The parameters passed from the client}

\item{ft3_cache}{The cachem cache (if relevant)}

\item{scratch_dir}{Where should the scratch files be written?}
}
\description{
Compile an Rmd flexTeaching assignment
}
