% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_filename.R
\name{ft3_create_filename}
\alias{ft3_create_filename}
\title{Title}
\usage{
ft3_create_filename(
  filename,
  assignment,
  id,
  seed,
  assignment_mode,
  solutions,
  file_content,
  ..
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
