% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assignments.R
\name{ft3_get_assignments_select2}
\alias{ft3_get_assignments_select2}
\title{Title}
\usage{
ft3_get_assignments_select2(
  assignments_dir = ft3_options("assignments_dir"),
  ...
)
}
\arguments{
\item{assignments_dir}{}

\item{...}{}
}
\description{
Title
}
