% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assignments.R
\name{ft3_get_assignments_simple}
\alias{ft3_get_assignments_simple}
\title{Title}
\usage{
ft3_get_assignments_simple(
  assignments_dir = ft3_options("assignments_dir"),
  current_assignment = NULL
)
}
\arguments{
\item{assignments_dir}{}
}
\description{
Title
}
