% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fingerprint.R
\name{ft3_get_fingerprint}
\alias{ft3_get_fingerprint}
\title{Title}
\usage{
ft3_get_fingerprint(seed, id, settings, assignment_mode)
}
\arguments{
\item{seed}{}

\item{id}{}

\item{assignment}{}
}
\description{
Title
}
