% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_error.R
\name{ft3_http_error}
\alias{ft3_http_error}
\title{Throw an HTTP error, with problem details}
\usage{
ft3_http_error(status_code, detail)
}
\arguments{
\item{detail}{}

\item{status}{valid}
}
\description{
Throw an HTTP error, with problem details
}
