% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_error.R
\name{ft3_http_error_wrapper}
\alias{ft3_http_error_wrapper}
\title{Title}
\usage{
ft3_http_error_wrapper(FUN, .custom_message = NULL, .status_code = 500, ...)
}
\arguments{
\item{FUN}{}

\item{.custom_message}{}

\item{...}{}
}
\description{
Title
}
