% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_options.R
\name{ft3_options}
\alias{ft3_options}
\title{Set or get options for my package}
\usage{
ft3_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{master_secret}}{(\code{character})  Master secret to use for generation of seeds. If NULL when the package is loaded, a random master seed is generated (with a warning). }
 \item{\code{scratch_dir}}{(\code{character})  Disk location to use a temporary space for rendering. If NULL, a temporary directory is used.}
 \item{\code{cache_location}}{(\code{character})  Disk location to save the content cache. If NULL, a temporary directory is used.}
 \item{\code{assignments_pkg}}{(\code{character}) Installed package to use for the assignments. If NULL, this package is used. }
 \item{\code{assignments_dir}}{(\code{character}) Directory to use for the assignments. If NULL, it will be determined from \code{assignments_pkg}.}
 \item{\code{default_file_icon}}{(\code{character}) Default icon to use for file downloads if none is specified. }
 \item{\code{timezone}}{(\code{character}) Default timezone to use for assignment restrictions. Empty string is system time zone. }
 \item{\code{lubridate_orders}}{(\code{character}) Time/date orders to use when parsing date-times using the lubridate package. }
 \item{\code{default_category_name}}{(\code{character}) The name of the category in which to place uncategorized assignments. }
 \item{\code{errors_to_client}}{(\code{logical}) Pass internal R error descriptions to the client interface? Doing this may pose a security risk, but may be helpful for debugging.}
}
}

