% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_alpha_num.R
\name{ft3_random_alpha_num}
\alias{ft3_random_alpha_num}
\title{Create random alphanumeric strings}
\usage{
ft3_random_alpha_num(n = 1, n_char = 10)
}
\arguments{
\item{n}{number of strings}

\item{n_char}{length of strings}
}
\description{
Create random alphanumeric strings
}
