% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_doc.R
\name{ft3_read_assignment_output}
\alias{ft3_read_assignment_output}
\title{Read the specified files that were written by an assignment}
\usage{
ft3_read_assignment_output(outfile, settings, pars)
}
\arguments{
\item{outfile}{Path to the file that was written}

\item{settings}{The compilation settings (from the yaml header)}

\item{pars}{The parameters passed from the client}
}
\description{
Read the specified files that were written by an assignment
}
