% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedStatusHook.R
\name{ft3_seed_status_hook}
\alias{ft3_seed_status_hook}
\title{knitr hook to reset seed after running chunk}
\usage{
ft3_seed_status_hook(before, options, envir, name)
}
\value{
Returns a hook function
}
\description{
This is automatically made available when .Rmd files are compiled.
}
\examples{

knitr::knit_hooks$set(ft3_seed_status = flexTeaching3.api::ft3_seed_status_hook)
}
