% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_api.R
\name{ft3_serve_api}
\alias{ft3_serve_api}
\title{Title}
\usage{
ft3_serve_api(
  cache_options = list(dir = ft3_options("cache_location"), destroy_on_finalize = TRUE),
  assignments_pkg = ft3_options("assignments_pkg"),
  assignments_dir = NULL,
  http_port = 8080,
  log_options = list(level = "off")
)
}
\arguments{
\item{assignments_pkg}{}

\item{http_port}{}

\item{log_options}{}
}
\description{
Title
}
