% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexTeaching3_R6.R
\name{FlexTeaching3}
\alias{FlexTeaching3}
\title{R6 Class Representing a set of flexTeaching3 assignments}
\description{
R6 Class Representing a set of flexTeaching3 assignments

R6 Class Representing a set of flexTeaching3 assignments
}
\examples{
# example code
ft3 <- FlexTeaching3$new()
rm(ft3)
Report settings in use by the FlexTeaching3 object

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FlexTeaching3-settings}{\code{FlexTeaching3$settings()}}
\item \href{#method-FlexTeaching3-new}{\code{FlexTeaching3$new()}}
\item \href{#method-FlexTeaching3-finalize}{\code{FlexTeaching3$finalize()}}
\item \href{#method-FlexTeaching3-clone}{\code{FlexTeaching3$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlexTeaching3-settings"></a>}}
\if{latex}{\out{\hypertarget{method-FlexTeaching3-settings}{}}}
\subsection{Method \code{settings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlexTeaching3$settings()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list containing the relevant settings
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlexTeaching3-new"></a>}}
\if{latex}{\out{\hypertarget{method-FlexTeaching3-new}{}}}
\subsection{Method \code{new()}}{
Create a new FlexTeaching3 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlexTeaching3$new(
  scratch_dir = NULL,
  assignment_pkg,
  master_seed,
  cachedisk_settings = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scratch_dir}}{Directory to use for creating/storing temporary files and cache. Defaults to a tempdir()}

\item{\code{assignment_pkg}}{Package containing flexTeaching3 assignment content}

\item{\code{master_seed}}{Master seed used for generating assignment seeds}

\item{\code{cachedisk_settings}}{an optional list of disk cache settings (see ?cachem::cache_disk)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new FlexTeaching3 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlexTeaching3-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-FlexTeaching3-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clean up when object is destroyed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlexTeaching3$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlexTeaching3-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FlexTeaching3-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlexTeaching3$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
