% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_skeleton.R
\name{ft3_assignment_skeleton}
\alias{ft3_assignment_skeleton}
\title{Title}
\usage{
ft3_assignment_skeleton(
  target_dir,
  title = "Example assignment",
  short_title = abbreviate(title),
  format = c("html_document", "html_fragment", "pdf_document"),
  ...
)
}
\arguments{
\item{target_dir}{}

\item{title}{}

\item{short_title}{}

\item{format}{}

\item{...}{Not currently used.}
}
\description{
Title
}
