% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_skeleton.R
\name{ft3_assignment_skeleton}
\alias{ft3_assignment_skeleton}
\title{Create a basic assignment}
\usage{
ft3_assignment_skeleton(
  target_dir,
  title = "Example assignment",
  format = c("html_document", "html_fragment", "pdf_document"),
  ...
)
}
\arguments{
\item{target_dir}{The directory that is to contain the assignments}

\item{title}{Title for the assignment}

\item{format}{One of 'html_document', 'html_fragment', or 'pdf_document'}

\item{...}{Not currently used.}
}
\value{
The path of the file written.
}
\description{
This function will create a folder and file that contains a basic assignment for use
with flexTeaching3. It is meant to be used on conjunction with developing a package of 
assignments; typically you'll point to the `inst/ft3_pkg` folder in the source of the 
assignments package you are developing.
}
\examples{
\dontrun{
ft3_assignment_skeleton('.', format = 'html_document')
}
}
