% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_filename.R
\name{ft3_create_filename}
\alias{ft3_create_filename}
\title{Create a file name from flexTeaching parameters}
\usage{
ft3_create_filename(
  filename,
  assignment,
  id,
  seed,
  assignment_mode,
  solutions,
  file_content,
  ...
)
}
\arguments{
\item{filename}{The original file name}

\item{assignment}{The assignment reference}

\item{id}{The user id parameter}

\item{seed}{The seed parameter}

\item{assignment_mode}{(logical) In assignment mode?}

\item{solutions}{(logical) Solutions requested?}

\item{file_content}{Full content of the file}

\item{...}{Other potential parameters (not currently used)}
}
\value{
A character string to be used as a file name
}
\description{
Create a file name from flexTeaching parameters
}
