% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_api.R
\name{ft3_serve_api}
\alias{ft3_serve_api}
\title{Start the flexTeaching3 API using RestRserve}
\usage{
ft3_serve_api(
  cache_options = list(dir = ft3_options("cache_location"), destroy_on_finalize = TRUE),
  assignments_pkg = ft3_options("assignments_pkg"),
  assignments_dir = NULL,
  http_port = 8080,
  backend = RestRserve::BackendRserve$new(),
  ...
)
}
\arguments{
\item{cache_options}{List of options to be passed to cachem::disk_cache()}

\item{assignments_pkg}{Package containing the assignments}

\item{assignments_dir}{Optional directory containing the assignments; if not specified, will be inferred from assignments_pkg.}

\item{http_port}{Port to start the server}

\item{backend}{RestRserve backend}

\item{...}{Further arguments to be passed to the RestRserve backend's start() method}
}
\value{
The result from the RestRserve backend's start() method
}
\description{
This function is meant to be run from a script, not in an interactive session.
See the file 'start_api.R' script included with this package.
}
